Make parallel rank and size per-MBCore-instance values, rather than global values.
Add MBProcConfig class for storing parallel rank and size, and as a place
to put utility methods for manipulating the processor id portion of the
entity id portion of MBEntityHandles.

Fix allocation of new entity handles (both sequences and individual handles)
such that they accept a processor id and olny allocate handles with that
processor id.


git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@924 6162379e-bd28-0410-9a7d-b7f4dcfcad3c