Fix portability issue storing MBMeshSet parent/child lists (fixes bug on Windows.)  Apparently if a bit-field value is signed, gcc treats the bits as the lower N bits of the declared type while Microsoft's compiler consideres them to be a signed integer of N bits.  Example:
struct S {
  int bits : 2;
};

int n = 2;
S s;
s.bits = n;
n = s.bits;

s.bits has the binary value 10b.  On Linux, n will finish with '2' while on 
Windows it is '-1'.  10b can be either, depending on whether or
one considers it to be a unsigned value or a 2-bit 2s-compliment value.

Many thanks to Byron Hanks for disconvery this issue, helping to track down
the cause, and for testing the fix.



git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@1393 6162379e-bd28-0410-9a7d-b7f4dcfcad3c