Free allocated arrays in any iGeom call that returns something other than iBase_SUCCESS (a few in iMesh_MOAB, many in iMeshP_MOAB).
Details:

Replace CHECK_SIZE macro with new macro: ALLOC_CHECK_ARRAY.  The argument
list for the new macro is a lot simpler: the type is inferred from
the array type and all array arguments (pointer, allocated, size) are
inferred from the array name. Thus:
  *adjacency_table_size = 16;
  CHECK_SIZE(*adjacency_table, *adjacency_table_allocated,
             *adjacency_table_size, int, iBase_MEMORY_ALLOCATION_FAILED);
becomes:
  ALLOC_CHECK_ARRAY(adjacency_table, 16);

However, the big difference is that if (and only if) ALLOC_CHECK_ARRAY 
allocated the array, then it will free it when the macro call goes out of 
scope (typically when returning from the function) unless KEEP_ARRAY(name) 
is called.  Thus one can safely write code like:

  ALLOC_CHECK_ARRAY(uv, 2*count);
  for (...) {
    ...
    if (error)
      RETURN(iBase_FAILURE);
    ...
    if (iBase_SUCCESS != *err)
      return;
  }
  KEEP_ARRAY(uv);
  RETURN(iBase_SUCCESS);
  
And even the hairy cases like the memory allocation for the second of
two arrays failing is handled correctly.

Also, provide the macro ALLOC_CHECK_ARRAY_NOFAIL for convenience.  It
is equivalent to consecutive calls to ALLOC_CHECK_ARRAY and KEEP_ARRAY.


git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@3570 6162379e-bd28-0410-9a7d-b7f4dcfcad3c