o Fix bug inserting entities into ranged (unordered) set. o Add test that reproduces conditions of above bug
This was interesting in that there was a fairly significant
bug in the insertion of entities into a range-compacted entity set
content list.  The bug resulted in range pairs not being correctly 
merged and not being kept in the expected sorted order, but all 
handles still beign represented.  This means that for the most common 
scenario of retreiving the contents of such a set in a Range, the 
results were always as expected.  It would have resulted in errors 
for many other cases, though, such as:
  o getting a subset by type or dimension might not return all entities
  o the contents retreived in a vector might contain duplicates
  o testing if a set contains a handle may return false negatives
  o issues with set boolean operations
  



git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@3999 6162379e-bd28-0410-9a7d-b7f4dcfcad3c