Fix a very old, but recently symptomatic, bug in DagMC.
When MOAB is built with CGM, the macro "CGM" is defined.  Until now,
this caused an extra data member to be included in the DagMC class.
However, client code compiling against DagMC does not define the CGM
macro (the macro is not exported to moab.make).  This means a client
compiling against DagMC.hpp can wind up with an incorrect map of the
data members in a DagMC object.

This hasn't surfaced as a problem before because the conditionally-
included data member was declared *after* all other client-facing
data members.  The reordering of data members in r4006 exposed the
problem.

The easy solution (which I've taken) is to simply remove the checks
from the header file.  If CGM is not available, the extra declarations
are harmless (and DagMC gets one extra unused std::vector).

Moral: Beware of #ifdefs in public headers, particularly if they
check for macros that client code may not know about.

git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@4016 6162379e-bd28-0410-9a7d-b7f4dcfcad3c