Now I handle check_valid flag to iterator creation, and perf test is restructured to test performance with and without check_valid.
One of the odd things about this is the timings I get:

100x100x100:
           Iterator   
   Test	   size	   !check_valid	    check_valid
e_to_v	   1	   .17		    .47
	   100	   .13		    .16

v_to_e	   1	   1.03		    .55
	   100	   .19		    .2	

200x200x200:
           Iterator   
   Test	   size	   !check_valid	    check_valid
e_to_v	   1	   2.03		    9.81
	   100	   1.83		    2.04

v_to_e	   1	   9.45		    10.93
	   100	   2.07		    2.35

The odd thing about the above is that the v_to_e query, for iterator size 1, is actually faster
when check_valid is true, then when it's false.  That's not an artifact, either, it's consistent 
across multiple runs.  Otherwise, checking validity seems to cost you about 10% overhead,
which is consistent with past observations.




git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@4857 6162379e-bd28-0410-9a7d-b7f4dcfcad3c