Set geom dimension and category tags on DagMC's implicit complement
This should make the complement volume act like a "real" volume
as it would be created by ReadCGM.  In particular, it allows
entity_by_id() to be called on the implicit complement, thus
fixing an odd corner case in the DagMC interface.

Update the expected volume counts in test_geom accordingly.

Also provide a DagMC function to test if a particular handle is the
implicit complement volume.

git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@4879 6162379e-bd28-0410-9a7d-b7f4dcfcad3c