fixes ticket 274
during communicating the shared sets information, from the remote
processors to the owning procesor, the set data is packed on the
remote proc, then sent to the ownig proc,
then unpacked on the owning processor, which will be in charge of
writing it to the file.
Unpacking means merging blocked ranges of file entities.
Instead of pointer logic, use Range, which is well tested for
merging subranges.
Also, if vectors need to be converted to compressed ranges, use
also the Range route, which is more stable