enhance crystal router example
add some options for the routing;
use ProgOptions class to parse input parameters
print timing information, and print tuples sent/received for one proc
only if requested
input the number of tuples sent from one task to a designated task
input the number of tasks each task will communicate with
the example could be used as a stress test and to measure communication
performance

the communication load is not balanced, some tasks receive/send more than
others; this is intentional, and part of the stress test

some receiving buffers will have to reallocate memory.

The reals sent are not verified, only the long data is checked for
what we expect to receive.