example to load partially a file, based on tagged sets
it is a good cherry-picking tool, when you want to extract only some
of the tagged sets in a h5m file. For example, you want to look only at the
mesh from partition 0, 1 and 4. You can use visit / subset capability.
Or you can use this example
It extends the option from mbconvert (-m), which can extract the blocks
for given material sets.
It will work for any tagged sets