fix the intersection on Equator
also, fix the interior to the CLL quad logic
tolerance for interior to the CS quad increased
to 2 times area tolerance.
will return 2 points if everything is on Equator (or 0?)
they might be collapsed later

still need to fix if a regular CL arc intersects a GC arc in
2 points.
for the current example, it is not needed, only if on Equator
there are 2 points in intx.

also, will need to qualify the edges in the polygons
some of them will be CL edges, and some will be GC arcs

Right now, everything is considered GC arcs in the polygons