3d tet refinement issue
for a particular model (tet mesh generated with
GenLargeMesh example, 3dsimplices.vtk), the nodes in the interior
have a pretty high valence (14 adjacent vertices,
14 adjacent edges, 36 adjacent triangles, and 24 adjacent tetras)
this model is a "complete" mesh, all faces and edges are
created

ahf uses a fixed limit for some local search array.
increasing this limit to 200 (from 100) avoids some
index overflow, and eliminates some valgrind-found illegal
writes and reads

with this fix, mbumr does not crash anymore on some buildbot builds
on gnep.
(the crash is happening when using:
./mbumr -d 3 -n 2 -s 3dsimplices.vtk )

for a more complex mesh, it would be possible that this max size is
still too low