Core::connect_iterate works with non congruently allocated cells of same type.
connect_iterate previously presumed that if a set of cell ids had
continuous ids, the allocation for those cells had to be continuous.

This is not the case if you manually allocate more than a single block
of cells of the same type, and between those allocations are any other
cell block allocation.

The solution to this problem is really simple, all we have to do is
use the min of the ElementSequence when it is smaller than the end iterator
we have been passed.