Several major optimizations for the intersection algorithm
1) Initial search for a seed

A while back, this was done with a kd tree.
But now, we do a linear search ( #$@%)
red cells are for sure covered by something, so we take a first red cell, and
find in (O(N)) time a seed among blue cells
it is faster than building a kd tree.
We do not need one

2) The problem is that we started before this change with a blue cell;
BIG MISTAKE.
There are many blue cells that are not covered, so it would be CRAZY slow
because we search linearly in all red cells

Additionally, we can squeeze more gains for moab intersections with some
additional memory usage and by storing the adjacent edges to each red cell
for a small mesh. The gain is about 15%, and is definitely worth it.

3) Replace bubble sort with std::sort for removing duplicates

It involves intersection points; order them according to angle from
their center of mass; it is not much faster than before, probably because number
of points is relatively low (10 on average)