add option for global visualization of any map
new option -o 2 (default) will create a mesh file in which
each weight from the map file is generating an edge from the
center of source to the center of target
Everything needed is read from the map file itself,
there is no need for source or target file, as for the other options
The centers of source cells are slightly elevated from the surface,
by a fraction set with option -r (fraction raise), default 0.05

This is just to identify better the source from the target;
Each source center has an "umbrella" for its weights

It offers some good insights og the maps used in E3SM

The main motivator for this kind of view is the rof2ocn map.
The pattern shown indicates how the targets (ocean cells) are
sometimes very far away from their sources.
This is why coverage approach failed to recover source cells
for some particular target cells